requirements.txt
streamlit
python-dotenv
Pillow
google-cloud-aiplatform
google-cloud-texttospeech
